==> 02-data.Rmd <==
# Data sources

```{r}
library(here)
```


The primary data source for this project is Global Health Data Exchange (GHDx) and [*World Bank*](https://data.worldbank.org/country). 

GHDx is a catalog of global health and demographic data. It provides statistics and other information about all kinds of diseases for different countries and for the globe as a whole. We use the [*GBD (Global Burden of Disease) Results Tool*](http://ghdx.healthdata.org/gbd-results-tool) built inside to generate our datasets about cancer in different aspects. The tool allows retrieval of GBD results and provides several variables for people to choose and combine. A brief introduction about the contents this tool supplies are as follows:

* __Base__: This menu provides 2 choices: *Single* means single year estimates, *Change* means multi-year estimates for select year ranges
* __Location__: This menu provides GBD super regions, regions, countries, subnational units, and custom regions (WHO regions, World Bank Income Levels, and more) to choose from
* __Year__: 1990-2019
* __Context__: This menu defines the type of indicator results people can access. It contains all GBD causes, risks, impairments, injuries by nature, etiology, etc. Once we select one of these results, a new drop-down menu about that result (varialbe) will appear and we can choose the values of our interest.
* __Age__: All GBD age groups
* __Metric(units)__: The choices are number, rate, and percent
* __Measure__: The choices are deaths, years of life lost (YLLs), years lived with disability (YLDs), disability-adjusted life years (DALYs), prevalence, incidence, life expectancy, healthy life expectancy (HALE), maternal mortality ratio (MMR), and summary exposure value (SEV)
* __Sex__: The choices are male, female, and both

If we generates the results for all measures for every single entry of the above menus, the dataset will be so huge and complicated since it contains data for countries all over the world for 30 years. So, to better conduct our analysis on several focuses, we generated three major datasets from GBD Results Tool and each group member was responsible for creating their dataset to solve corresponding questions. 

World Bank collects and provides large amount of global development data and was used as data supplement of countries' statistics in our dataset.

Details about the datasets are as follows:

## Data on Cancer by Risk

**Dataset: risk.csv**

The dataset is created from GBD Results Tool by choosing risk as context. 

```{r}
risk_desription <- readr::read_csv(here::here("./data/data_description.csv"))
knitr::kable(risk_desription[1:11, ], caption="Risk Data Desription", row.names=FALSE, font_size=10)
```

**Issues with this dataset: **

1. The column "location" only has one value "Global" so it is not a useful variable in the dataset. And the columns "upper" and "lower" give a boundaries to "val". In this project, we only used the "val" provided and not consider the range of "val". So, "upper" and "lower" columns were also not used. The unused columns need to be removed.

2. All the measures are assigned with three metrics: number, percent, and rate. Some metrics are not suitable to be used on some measures. For instance, percentage of years of life lost does not make much sense. So, we might consider deleting some meaningless or unused metrics to decrease data size.

3. The dataset is not tidy. Column "measure" actually contains the variables of our interests. And "val" stores the values of corresponding variable. This dataset needs to be pivot to a tidy format.


## Data on Cancer by Country Development

**Dataset: final.csv**

The dataset is combined with GBD results and the country dataset from World Bank.

```{r}
country_data<- readr::read_csv(here::here("./data/data_description.csv"))
knitr::kable(country_data[c(2,8,12,13,14,15,16,17,18), ], caption="Death among Countries Desription", row.names=FALSE, font_size=10)
```

**Issues with this dataset: **

1.There are only 19 countries in the dataset, which may not be supper sufficient for see the whole story on relationship between cancer death rate and country progress. 

3.The dataset are not large enough to show the main trend and death are affected by multiple reasons so the analyze results might include some bias.

## Data on Cancer over Time

**Dataset: making_progress.csv**
**Dataset: death_number.csv**

The dataset is created from GBD Results Tool by choosing cause as context. 

```{r}
making_progress_description<- readr::read_csv(here::here("./data/data_description.csv"))
make_progress_description<- making_progress_description[c(1,2,3,4,5,7,8,9,10,11),]
knitr::kable(make_progress_description[1:10,], caption="Making_progress_description & cmp Data Desription", row.names=FALSE, font_size=10)
```
**Issues with this dataset: **

1. Same as our topic 1, the column "location" only has one value "Global" which cannot provide more information. And the columns "upper" and "lower" give a boundaries to "val". However, we only used the "val" provided and not consider the range of "val". Also, the the column "sex" cares about both female and male. Hence, the unused columns need to be removed.

2. Considering that the dataset size need to be reduced, we may delete some meaningless and confusing metrics such as "YLLs percent" and "YLDs percent".

3. Since it take too long to download the whole dataset for different countries over 30 years. We downloaded two separated dataset with different "measure" and "age" group. That is why the two datasets have exactly the same data description.
