==> 03-cleaning.Rmd <==
# Data transformation

```{r}
library(readr)
library(tidyverse)
library(tidyr)
library(dplyr)
library(tidyr)
```

All of our datasets generated from GBD results tool need to be pre-processed as they are not in a tidy format originally. Please go to our [github repo chapter](https://github.com/Maggieli99/Cancerstatistic/blob/main/03-cleaning.Rmd) to see code in detail.

## Cancer Risk Data Transformation

For the risk data, we first removed the used columns "location", "upper" and "lower". And then we pivoted our dataset so that the combinations of *measure* and *metric* became new columns (variables) in the wider tidy data and values in *val* were placed in the corresponding entries. We also updated the columns names of our new data frame so that these names are more precise. Finally, noticed that all of our columns of character type have limited number of categories, we transformed all character columns into factors.

```{r}
cancer_risk <- read_csv("./data/risk.csv")
```
```{r}
# clean and pivot dataset
cancer_risk <- cancer_risk %>% 
  select(-c('location', 'upper', 'lower')) %>%
  pivot_wider(names_from = c(measure, metric), values_from = val) 

names(cancer_risk)[c(4, 9:14)]<-c('risk', 'YLDs_Number', 'YLDs_Percent', 'YLDs_Rate',  'YLLs_Number', 'YLLs_Percent', 'YLLs_Rate')
```

```{r}
# make str columns factors
cancer_risk$sex <- factor(cancer_risk$sex)
cancer_risk$age <- factor(cancer_risk$age)
cancer_risk$cause <- factor(cancer_risk$cause)
cancer_risk$risk <- factor(cancer_risk$risk)
cancer_risk$year <- factor(cancer_risk$year)
```

After data cleaning and transformation, our *cancer_risk* data frame has much less rows and contains only meaningful variables.

```{r}
str(cancer_risk)
```



## Cancer by Country Data Transformation

The original dataset of countries contains more than 2000 rows of basic information of each countries so here we pick important factors towards to death rate including GDP, GNI, health cost per capita and health cost percent of total GDP. Besides, we also spread the measure of cancer death including percent, rate and number. Here, we also add a year column. The updated dataset are summarized below and is tidy meanwhile.

```{r}
cancer = read.csv(file = './data/final.csv',header = TRUE)
cancer = cancer %>%
  select(-c('sex','age','upper','lower','cause')) 

##filter(cancer$year != '2000')
cancer = filter(cancer,cancer$measure == 'Deaths')
##cancer = filter(cancer, cancer$location != 'India')
country = read.csv(file = './data/allCountry.csv',header = TRUE)

country = country[which(country$Indicator.Code %in% c('NY.GNP.MKTP.PP.CD','NY.GDP.PCAP.PP.CD','SH.XPD.CHEX.PC.CD','SH.XPD.CHEX.GD.ZS')),]

country = country %>%
  select(c('Country.Name','Indicator.Name','X1990','X1995','X2000','X2005','X2010','X2015','X2019'))

country = country %>%
  gather(Year,Number,c(X1990,X1995,X2000,X2005,X2010,X2015,X2019))
country = country %>%
  spread(Indicator.Name,Number)

country = country %>% mutate(Year=recode(Year, 
                         'X1990'="1990",
                         'X1995'="1995",'X2000'="2000",'X2005'="2005",'X2010'="2010",'X2015'="2015",'X2019'="2019"))

colnames(country) = c('country','year','Health_Cost_of_GDP','Health_Cost','GDP','GNI')
country$year = as.factor(country$year)
cancer$year = as.factor(cancer$year)
cancer$location = as.factor(cancer$location)
country$country = as.factor(country$country)
levels(cancer$location)[levels(cancer$location)=="United States of America"] <- "United States of America"
levels(country$country)[levels(country$country)=="United States"] <- "United States of America"
levels(cancer$location)[levels(cancer$location)=="Australasia"] <- "Australia"
levels(country$country)[levels(country$country)=="Egypt, Arab Rep."] <- "Egypt"
levels(country$country)[levels(country$country)=="Congo, Rep."] <- "Congo"
levels(country$country)[levels(country$country)=="Iran, Islamic Rep."] <- "Iran"
levels(cancer$location)[levels(cancer$location)=="Iran (Islamic Republic of)"] <- "Iran"
data = full_join(cancer,country, by = c('location' = 'country','year' = 'year'))
data = data %>%
  spread(metric,val)
data = data[-c(1)]
data$year  = as.character(data$year)
cat(NULL,file="./data/finalData.csv")
write_csv(data,'./data/finalData.csv')
str(data)
```


```{r}
cancer = read.csv(file = './data/finalData.csv',header = TRUE)
cancer = cancer %>% select(c(year,Rate,location)) %>%
  spread(year,Rate)
cat(NULL,file="./data/interactive.csv")
write_csv(cancer,'./data/interactive.csv')
```


## Cancer Time Series Data Transformation

For the making_progress data, we need to do the similar step with risk data, that is, we first removed the used columns "location", "upper" and "lower","sex", since what we need to show is a global trend for both female and male. And then we pivoted our dataset so that the combinations of *measure* and *metric* became new columns (variables) in the wider tidy data and values in *val* were placed in the corresponding entries. To be more precise and clear, we need to change the column name for the metric. 

```{r}
cancer_making_progress <- read_csv("./data/making_progess.csv")
```

```{r}
# clean and pivot dataset
cancer_making_progress <- cancer_making_progress %>% 
  select(-c('location', 'upper', 'lower', 'sex')) %>%
  pivot_wider(names_from = c(measure, metric), values_from = val) 

```

```{r}
names(cancer_making_progress)[c(4,6,7,8)]<-c('DALYs_Rate', 'YLLs_Rate', 'YLDs_Percent', 'YLDs_Rate')
```

```{r}
cmp<- read_csv("./data/death_number.csv")
cmp <- cmp %>% 
  select(-c('location', 'upper', 'lower', 'sex')) %>%
  pivot_wider(names_from = c(measure, metric), values_from = val) 

```

```{r}
str(cancer_making_progress)
```

```{r}
str(cmp)
```
