==> 05-results.Rmd <==
# Results

```{r}
library(tidyverse)
library(tidyr)
library(dplyr)
library(highcharter)
library(ggplot2)
library(ggmosaic)
library(treemap)
library(plotly)
library(ggpubr)
library(RColorBrewer)
library(ggthemes)
```


## The Effect of External Factors on Cancer

The etiology of cancer is very complicated and has not been fully understood. But it has been found that the two major causes of cancer are genetic factors and external factors. While one of the biggest fears of human comes from cancer and people usually feel helpless in the face of cancer, we should realize that keep good life style might significantly reduce our chance of getting some kind of cancer. So below we analyze the effect of external factors on cancer and give potential ways of preventing cancer.


### Share of cancer deaths attributed to risk factors from 1990-2019

```{r}
risk_level1 <- c("Environmental/occupational risks", "Behavioral risks", "Metabolic risks")
cancer_risk_1 <- cancer_risk %>% 
  filter(sex=="Both", risk %in% risk_level1) %>%
  group_by(caus, year) %>%
  mutate(share_of_death = sum(Dt_N*Dt_P)/sum(Dt_N)) %>% 
  summarize(share_of_death=mean(share_of_death))
```  

```{r}
series <- cancer_risk_1 %>% 
  group_by(caus) %>% 
  do(item = list(
    cause = first(.$caus),
    sequence = .$share_of_death,
    value = first(.$share_of_death)
  )) %>% 
  .$item

series <- series[order(sapply(series,"[[",3),decreasing = TRUE)]
```


```{r}
highchart() %>%
  #hc_chart(data = series, type="bar", hcaes(y = sequence, color = cause)) %>%
  hc_add_series(data = series, type="bar") %>% 
  hc_motion(
    enabled = TRUE,
    labels = cancer_risk_1$year,
    axisLabel = "year",
    #series = c(0,1),
    autoplay=TRUE
  ) %>%
  hc_xAxis(categories = sapply(series,"[[",1)) %>%
  hc_yAxis(min = 0, max = 1)%>%
  hc_colorAxis(stops = color_stops(n = 3, colors = c("firebrick2","white","dodgerblue3")))%>%
  hc_title(text = "<b>Share of cancer deaths attributed to risk factors, 1990-2019</b>") %>%
  hc_legend(align="right")

```

The plot above shows an estimation of share of global cancer deaths which are attributed to external risk factors over the 30 years. For example, around 71% of larynx cancer deaths are attributed to some external risk. Only the remaining 29% of deaths are assumed to have no attribution to risk factors, and would occur naturally in the absence of external factors. So, we can see that the cancers with blue bar, like cervical cancer and mesothelioma, are more caused by external risks; the cancers with red bar, like thyroid cancer and ovarian cancer, are more due to genetic factors and so that people can do less to prevent such cancers. 

We can notice that cervical cancer is 100% due to external risks according to the plot. It might be too extreme to say but it is a suggestion that people can definitely do a lot of things to successfully prevent having such cancer.

We also observe that over the past 30 years, though the overall share of cancer deaths due to external risks is not changing dramatically, the external risk attribution does have some changes for some kind of cancer. For instance, esophageal cancer is less and less due to external risks. This may be because the food people nowadays are more various and cleaner and people take drinking water more seriously than before. So people are having less chancing getting esophageal cancer due to lack of vegetable eating and water drinking. On the contrary, liver cancer are more and more due to external risks. This may have relation with the constantly increasing smoke rate of human.

### Risk distribution over different types of cancer (1990, 2019)

```{r}
cancer_risk_2 <- cancer_risk %>% 
  filter(sex=="Both", risk %in% risk_level1) %>%
  group_by(year, caus,risk) %>%
  mutate(Freq = sum(Dt_N)) %>% 
  summarize(Freq=mean(Freq))
```

```{r}
data_1990 <- cancer_risk_2%>%filter(year==1990)
data_1990$risk <- droplevels(data_1990)$risk
```

```{r}
a<-ggplot(data = data_1990) +
  geom_mosaic(aes(weight=Freq,x=product(risk), fill = risk, 
                  conds = product(caus)), show.legend = FALSE) +
  labs(title='Risk Proportions in 1990',x = "Cancer") +
  theme(axis.text.x = element_text(angle = 90),axis.text=element_blank(),text = element_text(size=8))
```

```{r}
data_2019 <- cancer_risk_2%>%filter(year==2019)
data_2019$risk <- droplevels(data_2019)$risk
```

```{r}
b<-ggplot(data = data_2019) +
  geom_mosaic(aes(weight=Freq,x=product(risk), fill = risk, 
                  conds = product(caus))) +
  labs(title='Risk Proportions in 2019',x = "Cancer") +
  theme(axis.text.x = element_text(angle = 90),axis.text=element_blank(),axis.ticks.y=element_blank(),axis.title.y = element_blank(),legend.text=element_text(size=7),text = element_text(size=8))
```

```{r}
a+b
```

In the first plot, we can only see external risks as a whole. So, in the second plot above, we further explore what exactly those risks are and how people can do to prevent specific kind of cancer potentially. GHDx defines three major external risks: behavioral risks, environmental/occupational risks, and metabolic risks. From the mosaic plot above, we can see that behavioral risks are the biggest group of risks, and environmental risks are relatively the smallest group. Thus, there is a lot of things human can control to prevent them from having cancer.

None of these cancers have a very even distribution of these three risks. They all bias to some one or two kinds of risks, which correspond to their features and give us suggestions on how we can fight against these cancers.

We notice that cervical cancer, pancreatic cancer and prostate cancer are almost all due to behavioral risks. That means, by controlling our behavior or actively making some changes to our life style, the risk of developing these cancers can be minimized at most. Gallbladder and biliary tract cancer, multiple myeloma, non-hodgkin lymphoma, and uterine cancer are almost all due to metabolic risks. Metabolic risks are basically about physique of people. These risks have already built into the human body. So a better way to guard against these cancer might be to properly deal with these physical problems to avoiding them from hurting their bodies further. Finally, we can see that mesothelioma is almost all due to environmental/occupational risks. It makes sense because we know that it is mostly caused by exposure to asbestos. If people does not live in an environment with high asbestos density, then risk of getting mesothelioma is supper low. For people have to exposure to asbestos, protection is especially important.

Comparing the risk proportion from 1990 to 2019, we can still see that there is not much difference. But we can indeed see that the proportion of metabolic risks increased and the proportion of behavioral risks decreased by a little. This change may indicate human have improved their behaviors to make progress against cancer, but in the mean while, increasing pressure also cause more other physical and metabolic problems that lead to getting cancer.


### Death Number and Rate by Risks over the 30 years (1990-2019)

```{r}
bahavioral_risk <- c("Alcohol use", "Dietary risks", "Drug use", "Low physical activity", "Tobacco", "Unsafe sex")
environmental_risk <- c("Air pollution", "Occupational risks", "Other environmental risks")
metabolic_risk <- c("High fasting plasma glucose", "High body-mass index")

cancer_risk_3 <- cancer_risk %>% 
  filter(sex=="Both", !risk %in% risk_level1) %>%
  group_by(caus, risk) %>%
  mutate(risk_category=case_when(risk %in% bahavioral_risk ~"Behavioral",
         risk %in% environmental_risk ~"Environmental",
         risk %in% metabolic_risk ~"Metabolic"),
         death_rate=sum(Dt_R),
         death_number=sum(Dt_N)) %>%
  summarize(risk_category=first(risk_category), death_number=mean(death_number),death_rate=mean(death_rate))
```

```{r}
treemap(cancer_risk_3, 
        index=c("risk_category", "risk"),
        vSize="death_number",
        vColor="death_rate",
        type="value",
        format.legend = list(scientific=FALSE, big.mark=" "))
```

When talking about behavioral risks or metabolic risks, it is still a little bit abstract and general. So, to get an even better understanding of all these risks for cancer, we sum up the data for every year and created a tree map to break down the risk categories into even deeper levels. In the above tree map, the area means the average number of people death from cancer because of the risk stated per year, the color means death rate (number of deaths per 100k people). We can clearly see that tobacco is definitely the most common and dangerous carcinogen. It leads to the largest number of people to die from cancer and has the largest death rate as well. 
For risks from environment, air pollution and occupational risks are the two major components. So people might want to wear face masks and make good protection against the environmental attack. For metabolic risks, high body-mass index and high fasting plasma glucose are the two main causes. That means people with obesity and high blood sugar need special attention for cancer prevention.


### Death Rate by Sex and Age for Different Risks

```{r}
cancer_risk_4 <- cancer_risk %>% 
  filter(sex!="Both", risk %in% risk_level1) %>%
  group_by(sex, age, risk, year) %>%
  mutate(death_rate=sum(Dt_R)) %>%
  summarize(death_rate=first(death_rate))
```

```{r}
behavioral_death <- cancer_risk_4 %>% filter(risk=="Behavioral risks")
environment_death <- cancer_risk_4 %>% filter(risk=="Environmental/occupational risks")
metabolic_death <- cancer_risk_4 %>% filter(risk=="Metabolic risks")
```

```{r}
g1<-ggplot(behavioral_death, aes(x=year, y=death_rate, group=sex:age))+
  geom_line(aes(color=sex))+
  geom_point(aes(shape=age),size=1,show.legend = FALSE)+
  labs(title='Death due to Behavioral Risk by Age & Sex', y="Death Rate\n per 100k") +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        axis.ticks.x = element_blank())

g2<-ggplot(environment_death, aes(x=year, y=death_rate, group=sex:age))+
  geom_line(aes(color=sex),show.legend = FALSE)+
  geom_point(aes(shape=age),size=1)+
  labs(title='Death due to Environmental Risk by Age & Sex', y="Death Rate\n per 100k") +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        axis.ticks.x = element_blank())

g3<-ggplot(metabolic_death, aes(x=year, y=death_rate, group=sex:age))+
  geom_line(aes(color=sex),show.legend = FALSE)+
  geom_point(aes(shape=age),size=1,show.legend = FALSE)+
  labs(title='Death due to Metabolic Risk by Age & Sex',y = "Death Rate\n per 100k") +
  theme(axis.text.x = element_text(angle = 90))

g1 + g2 + g3 + plot_layout(ncol=1)
```

For this plot, we again pay attention to the three main risk categories. But this time, we focus on people of different sex and age because the cancers that men and women have are very different. First we can see that for babies and kids (age 0 to 14), the death rate of all the risks are 0. That makes sense since kids and babies are very unlikely to have cancer. Even if they have, they have a longer year of life with disability and are extremely unlikely to die before 14 years old. And the same situation holds for teenagers and people of prime. It shows that cancer favors older people.  

For age group form 50 to 74 years old, we now can see some variations among male and female and risks. For metabolic risk, the death rates of male and female are almost the same with males' slightly higher. As for environmental risk, women's death rate is very close to 0 while men's death rate is a little more significantly higher. For behavioral risk, the difference between men and women is even more significant.

And this kind of gap is even more obvious among 75 plus elders. Except for metabolic risk where the death rate of both mean are women are much higher than other age groups, death rate of elder women is almost the same as that of 50 to 74 age group, but death rate of elder men is dramatically higher. It indicate that among elders, cancer favors elder men. So men should be more careful about these risks, especially behavioral risk, to reduce their probability of getting cancer, for example, they should think of quit smoking, each more vegeables, and drink less alcohol.

Over the 30 years, environmental risks has always been the leading role of causing cancer, death rate came from metabolic risk increases, while death rate came from environmental risk and behavioral risk started to decrease since 2008. These results all further confirm the conclusion I got from the previous plots.




## Relationship between Cancer Deaths and Country Development

### Heatmap of Death Rate in World

The heatmap of death rate can show the death rate among countrie from 1990 to 2019. You can also click the run button and then we also see the death rate change from 1990 to 2019. From 1990 to 2019, the red area where death rate become higher so the color is lighter. However, from 1990 to 2019, the death rate in blue zone where death rate are high become lower when the color are lighter. From the graph, wee can notice that in some developed countries including US and UK, the death rate is higher than other developing countries all the time. Even though, the death rate is lower from 1990 too 2015 in US, it still higher than other developing country.  

```{r}
data = read_csv(file = './data/finalData.csv')
new_data = data %>% 
  group_by(location) %>%
  do(item = list(country = first(.$location),sequence = .$Rate), value = first(.$Rate))%>%
    .$item

highchart(type = 'map') %>%
  hc_add_series(data =  new_data,joinBy = c('name','country'), mapData = worldgeojson, dataLabels = list(enabled = TRUE, format = "{point.properties.postalcode}"))  %>%
  hc_colorAxis(stops = color_stops(n = 3, colors = c("firebrick2","white","dodgerblue3")),
               max=400,min=0)  %>%

  hc_motion(enabled = TRUE,
    axisLabel = "year",labels =as.character(sort(unique(data$year),decreasing = FALSE)),
    magnet = list(
      round = "floor" 
    ))%>% 
  hc_legend(layout = "vertical", 
            align = "right",
            title = list(text = 'Death Rate')) %>%
  hc_add_theme(hc_theme_smpl()) %>%
  hc_exporting(enabled = TRUE)  %>%
  hc_title(text = "Death Rate In The World From 1990 To 2019")

```

### Bubble Plot of Death Rate vs Health Cost of GDP

The bubble plot can show the death rate versus health cost of GDP. The size of bubble can also show the GDP per capita. If you move the mouse over the bubble, we can also see the name of country. The result is interesting: in most cases, higher health cost of GDP means the high death rate and higher GDP. However, America does not follow this rule: America have high health cost of GDP but the death only in the middle level. The rules are followed across the years. To sum up, higher income and health expenditure can not promise the lower risk and death of cancer, because in high-income countries, the other risk brought by technology can cause the higher death rate of cancer.  

```{r}
new_data = data[-which(data$location=='Brazil'),]
new_data = new_data[new_data$year %in% c('2000','2005','2010','2015'),]
x2000 = new_data[new_data$year=='2000',]
x2005 = new_data[new_data$year=='2005',]
x2010 = new_data[new_data$year=='2010',]
x2015 = new_data[new_data$year=='2015',]
```

```{r}
new_data
fig <- new_data %>%
  plot_ly(
    x = ~Health_Cost_of_GDP, 
    y = ~Rate, 
    size = ~GDP, 
    color = ~Rate,
    frame = ~year, 
    text = ~location, 
    hoverinfo = "text",
    type = 'scatter',
    mode = 'markers'
  )%>% animation_opts(
    1000, easing = "elastic", redraw = FALSE
  )
fig
```

### Accumulated Death Among Countries



In this plot, I stacked the total amount of death from 2000 to 2015. Here, it is easy to see that China has much larger amount of death than other countries. The death amount is hugely depend on the population of each countries, so many countries with small population has small death amount and countries with large population has large death amount. Interestingly, even though India has large population and almost the same population , the death amount is still much smaller than China. Another interesting fact is that although US has only about 1/4 population as India, the death amount is almoost the same. Therefore, people should pay more attention about thee information behind the death amount graph. 



```{r}
COLORS <- RColorBrewer::brewer.pal(6, "Set2")

new_data %>%
  hchart('column',hcaes(x = 'location', y = 'Number',group = 'year'),stacking="year") %>%
  hc_legend(title = list(text = "Year",style=list(fontSize='10px')),
            align = "right", verticalAlign = "middle",layout = "vertical",floating=FALSE) %>%
  hc_colors(COLORS)%>%
  hc_chart(zoomType = "x") %>%
  hc_xAxis(title = list(text = "Country")) %>%
  hc_yAxis(title = list(text = 'Accumulated Death'))  %>%
  hc_subtitle(text = "Click on the legend to show selected years") %>%
  hc_exporting(enabled = TRUE)
```

### Death Rate by GDP



Here, I combine four plots of death rate by GDP from 2000 to 2015. In this plot, I only try to explore the death rate affected by GDP. The four plots show that the trend are not changed from 2000 to 2019: high GDP means high death rate. Even though high GDP mean high income and better living condition, it does not mean the death rate are lower. In highly developed countries, there are still many factors that can affect people's health and cancer is still a big threaten to developed countries. 



```{r}
new_data  =  data %>%
  mutate(scaled_GDP = GDP/100)

new_data = new_data[-which(new_data$location=='Brazil'),]
new_data = new_data[new_data$year %in% c('2000','2005','2010','2015'),]
x2000 = new_data[new_data$year=='2000',]
x2005 = new_data[new_data$year=='2005',]
x2010 = new_data[new_data$year=='2010',]
x2015 = new_data[new_data$year=='2015',]
plot_2000 = ggplot(data = x2000,aes(x = scaled_GDP, y = Rate))+geom_line(colour="#FF9999")+geom_point(colour="#FF9999")+ggtitle("Death Rate by GDP in 2000") +
  labs(x ="GDP", y ="Death Rate")+
  theme_grey(10)+
  theme(plot.title =element_text(face='bold', color ='steelblue2'))

plot_2005 = ggplot(data = x2005,aes(x = scaled_GDP, y = Rate))+geom_line( colour="#E69F00")+geom_point(colour="#E69F00")+ggtitle("Death Rate by GDP in 2005") +
  labs(x ="GDP", y ="Death Rate")+
  theme(plot.title =element_text(face='bold', color ='steelblue2'))

plot_2010= ggplot(data = x2010,aes(x = scaled_GDP, y = Rate))+geom_line(colour="#F0E442")+geom_point(colour="#F0E442")+ggtitle("Death Rate by GDP in 2010") +
  labs(x ="GDP", y ="Death Rate")+
  theme(plot.title =element_text(face='bold', color ='steelblue2'))

plot_2015 = ggplot(data = x2015,aes(x = scaled_GDP, y = Rate))+geom_line(colour="#CC79A7")+geom_point(colour="#CC79A7")+ggtitle("Death Rate by GDP in 2015") +
  labs(x ="GDP", y ="Death Rate")+
  theme(plot.title =element_text(face='bold', color ='steelblue2'))

ggarrange(plot_2000,plot_2005,plot_2010,plot_2015, 
          label.y = 1,
          align='hv',
          ncol = 1, nrow = 4)
```

## Is the World Making Progress against Cancers?

### Age-standardized cancer death rates

```{r}
cause_type <- c("Stomach cancer", "Breast cancer", "Liver cancer", "Tracheal, bronchus, and lung cancer", "Nasopharynx cancer", "Esophageal cancer", "Kidney cancer", "Mesothelioma")
cancer_making_progress1 <- cancer_making_progress[,c(2,3,5)] %>% 
  filter(cause %in% cause_type) %>%
  group_by(cause, year)%>%  
  mutate(Death_rate=first(Deaths_Rate))%>% 
  summarise(Death_rate=first(Deaths_Rate))
```  


```{r}
ggplot(data = cancer_making_progress1, aes(x=year,y=Death_rate, group =cause ,color=cause,shape=cause))+
  geom_point()+
  geom_line()+
  xlab("Year")+ 
  ylab("Deaths Rate")+
  ggtitle("The Death Rate from 1990 to 2019 by Cancer Type")+
  theme_bw() +
  theme(panel.grid.major=element_line(colour=NA),
        panel.background = element_rect(fill = "transparent",colour = NA),
        plot.background = element_rect(fill = "transparent",colour = NA),
        panel.grid.minor = element_blank(),
       legend.key.size = unit(2, 'mm'),
       axis.text.x = element_text(angle=90),
        legend.box.background = element_rect(color="black"))+
  scale_x_continuous(limits = c(1990,2019),breaks = seq(1990,2019,1))


```



We randomly chose eight common cancers in our life to watch how their death_rate change in thirty years. We cannot only look at absolute death numbers to see if we've made success in the fight against cancer. It doesn't explain the world's population expansion. Hence, health statisticians investigate death rate, which calculates the number of cancer deaths per 100,000 individuals. It is conspicuous that the death rate shows a decreasing trend in those 30 years especially for Liver cancer, Esophageal cancer, stomach cancer and Tracheal, bronchus and lung cancer. Moreover, this measure is based on Age-Standardized rate, which tells us what the death rate will be if the age structure of the population remains unchanged over time. Hence, we can say the world is making slow progress. Most deaths are caused by smoking since we can observe Tracheal, bronchus and lung cancer contributes to the highest death rate globally. Here is a strong decline of death rates for lunger cancer since smoking rate is decreasing in these years.


### The number of cancer deaths is increasing for all age

```{r}
cause_type <- c("Stomach cancer", "Breast cancer", "Liver cancer", "Tracheal, bronchus, and lung cancer", "Nasopharynx cancer", "Esophageal cancer", "Kidney cancer", "Mesothelioma")
cancer_making_p <- cmp[,c(2,3,4)] %>% 
  filter(cause %in% cause_type) %>%
  group_by(cause, year)%>% 
  mutate(Death_numbers=first(Deaths_Number))%>% 
  summarise(Death_numbers=first(Deaths_Number))
```

```{r}
options(scipen=200)
ggplot(cancer_making_p,aes(year,Death_numbers,fill=cause))+
geom_area(position="stack")+
ggtitle("Number of Cancer Deaths by Type from 1990 to 2019")+
scale_fill_economist()+
guides(fill=guide_legend(title=NULL)) +
theme( panel.grid.major=element_line(colour=NA),
        plot.background = element_rect(fill = "transparent",colour = NA),
  legend.text = element_text(size=8), axis.text.x = element_text(angle=90))+
scale_x_continuous(limits = c(1990,2019),breaks = seq(1990,2019,1))
```


The number of cancer deaths grew from roughly about 3 million in 1990 to 5 million in 2019 for those eight random cancers, as indicated in the graph.

There are two causes for this, as we've just seen: The first is that the global population is growing, which is followed by an increase in the annual death toll.

The second major reason is that the world has made significant progress in decreasing causes of mortality that used to kill individuals young, particularly infectious illnesses.This indicates that the world's population is aging, and an increasing number of individuals are dying of cancer and other causes.

We may predict a rise in the number of individuals dying early as the globe grows wealthier.


The number of cancer deaths grew from roughly about 3 million in 1990 to 5 million in 2019 for those eight random cancers, as indicated in the graph. Evidently, tracheal, bronchus, and lung cancer contributes most to the death numbers for "All age". As we discussed before, it also plays an critical role in age-standardized death rate. Stomach cancer is the second big reason that lead to the death globally. Mesothelioma is the least cancer that result in death. 
From what the figure shown, we found there is a obvious increase in death number. There are two causes for this the first is that the global population is growing, which is followed by an increase in the annual death toll.
The second major reason is that the world has made significant progress in decreasing causes of death that can be detected at an individual's young age. Since the development of technology such as the creation of advanced instruments and the doctor's outstanding medical skills, some disease may not develop to cancer. However, there is still an increase in death number; that is, the population of the world is aging and this means when people are getting old, they will die because of cancer. Also, this reason is a natural occurrences when people getting old. 
Also, we can predict that the death number would continuously increase since diseases may not transfer to cancer because of the prevention by advanced medical technology.



### Year of life lost rate is decreasing.

```{r}
cause_type <- c("Stomach cancer", "Breast cancer", "Liver cancer", "Tracheal, bronchus, and lung cancer", "Nasopharynx cancer", "Esophageal cancer", "Kidney cancer", "Mesothelioma")
cmp2 <- cancer_making_progress[,c(2,3,6)] %>% 
  filter(cause %in% cause_type) %>%
  group_by(cause,year) %>% 
mutate(Year_of_Life_Lost=first(YLLs_Rate))%>% 
summarise(Year_of_Life_Lost=first(YLLs_Rate))
```


```{r}
series <- cmp2 %>% 
  group_by(cause) %>% 
  do(item = list(
    cause = first(.$cause),
    sequence = .$Year_of_Life_Lost,
    value = first(.$Year_of_Life_Lost)
  )) %>% 
  .$item

series <- series[order(sapply(series,"[[",3),decreasing = TRUE)]
```


```{r}
highchart() %>%
  hc_title(text = "Year of Life Lost Rate in 30 years ") %>%
  hc_add_series(data = series, type="bar") %>% 
  hc_legend(align = "right") %>%
  hc_motion(
    enabled = TRUE,
    labels = cmp2$year,
    axisLabel = "year",
    #series = c(0,1),
    autoplay=TRUE
  ) %>%
  hc_yAxis(title = list(text = "Year of Life Lost Rate")) %>%
  hc_xAxis(title = list(text = "Cancer Type")) %>%
  hc_xAxis(categories = sapply(series,"[[",1)) %>%
  hc_yAxis(min = 0, max = 700)%>%
  hc_colorAxis(stops = color_stops(n = 3, colors = c("pink","red","purple")))

```



The plot above shows an estimation on cancer patients' years of life lost rate from 1990 to 2019 based on eight random cancers. By dragging the time line, we can easily find that there is a big decrease of the year of life lost rate. Also, the rate is calculated by adding the year of life lost rate for 100k people that got this cancer. The decrease in year of life lost rate means people who got thee cancer can live longer than that 30 years before. As we can seen the purple line which is the year of life lost rate for people who got tracheal, bronchus, and lung cancer decrease from 651 to 544, that is decreased by 16.4%. What's more the most red line which represents the year of life lost rate for people who got stomach cancer decrease from 487 to 264, which is decreased by 46%. This is a huge change. I believe the reason leads to the huge change must rely on the world's making progress. For example, patients can be treated at once. People get richer and they can afford the fee to do the operations if they were detected early-stage cancer.




### The tend of rate of years for who is living with disability is flat or increase a little. 

```{r}
cause_type <- c("Stomach cancer", "Breast cancer", "Liver cancer", "Tracheal, bronchus, and lung cancer", "Nasopharynx cancer", "Esophageal cancer", "Kidney cancer", "Mesothelioma")
cmp3 <- cancer_making_progress[,c(2,3,8)] %>% 
  filter(cause %in% cause_type) %>%
  group_by(cause,year) %>% 
mutate(Year_of_Life_with_Disability=first(YLDs_Rate))%>% 
summarise(Year_of_Life_with_Disability=first(YLDs_Rate))
```

```{r}
ggplot(data = cmp3, aes(x=year,y=Year_of_Life_with_Disability, group =cause ,color=cause,shape=cause))+
  geom_point()+
  geom_line()+
  xlab("Year")+ 
  ylab("Year of Life with Disability")+
  ggtitle("The Year Living with Disablity from 1990 to 2019")+
  theme_bw() +
  theme(panel.grid.major=element_line(colour=NA),
        panel.background = element_rect(fill = "transparent",colour = NA),
        plot.background = element_rect(fill = "transparent",colour = NA),
        panel.grid.minor = element_blank(),
       legend.key.size = unit(2, 'mm'),
       axis.text.x = element_text(angle=90),
        legend.box.background = element_rect(color="black"))+
  scale_x_continuous(limits = c(1990,2019),breaks = seq(1990,2019,1))
```

The plot above shows an estimation on cancer patients' years of living with disability from 1990 to 2019 based on eight random cancers. Also, the rate is calculated by adding the year of living with disability rate for 100k people that got this cancer. The red line which represents the breast cancer shows a little bit increase, which means people who lived with breast cancer live longer now than 30 years before. Also, the green one with rectangle shape and the blue one with snow shape increase a little bit as well. They represent people's living year with Kidney cancer and Nasopharynx cancer. However, other line such as the purple one which represents the Stomach cancer surprising is decreasing. The little increase may related to the world progress on the advanced medical equipment as we discussed before. Also, breast cancer usually be detected among women. A good mood and an optimistic attitude is the most factor that can live longer. What's more, the society is more likely to value equality than that before and women plays a more important role in this world which can also be the reason. For the decrease on Stomach cancer, I think the reason is that unhealthy instant food is filled in people's life. Also, keeping staying up late becomes the norm among people. Hence, cancer patients' years of living with these disability a little bit decrease.
