==> 06-interactive.Rmd <==
# Interactive component

<html lang="en">
<head>
  <meta charset="UTF-8">
  <script src="https://d3js.org/d3.v7.js"></script>
  <title>Interactive Plot</title>
</head>

<body>
  <header>
    <h2><b>Death rate (per 100k) by Country over Years</b></h2>
  </header>
  
  <p>
  <b>Instructions:</b>
  
    1. Click the buttons to compare death rate in different years
    
    2. Hover the mouse over a bar (bar turns orange) to see the exact death rate of a country
    
  <b>Description:</b>
  
    The interactive plot is built with D3 version 7. It shows the death rate of 19 chosen countries from year 1990 to year 2019 with every 5 years apart (except for 2019 since we do not have data for 2020). By clicking the "year" button in order, we can clearly see how and how fast the death rate increases over years. If look carefully, we can notice that the death rate for Japan always increases at a fast rate. It is surprising since people may think developed country should have excellent medical facility and so lead to slow increasing in cancer death rate. But the plot shows that developed countries actually have larger increasing rate of cancer death rate than that of developing and under developed country.
    
  </p>


 <div id="interaction">
  <button onclick="update('1990')">Year 1990</button>
  <button onclick="update('1995')">Year 1995</button>
  <button onclick="update('2000')">Year 2000</button>
  <button onclick="update('2005')">Year 2005</button>
  <button onclick="update('2010')">Year 2010</button>
  <button onclick="update('2015')">Year 2015</button>
  <button onclick="update('2019')">Year 2019</button>
 </div>
 
 
 <script>
   var margin = {top: 30, right: 30, bottom: 200, left: 50};
   var height = 600 - margin.top - margin.bottom;
   var width = 800 - margin.right - margin.left;
   var svg = d3.select("#interaction")
                .append("svg")
                  .attr("width", width + margin.left + margin.right)
                  .attr("height", height + margin.top + margin.bottom)
                .append("g")
                  .attr("transform","translate(" + margin.left + "," + margin.top + ")");

   var x = d3.scaleBand().range([0,width]).paddingInner(0.05);
   var y = d3.scaleLinear().range([height,0]);
   var xAxis = svg.append("g")
     .attr("transform", "translate(0," + height + ")");
   var yAxis = svg.append("g")
     .attr("class", "myYaxis1");

   function update(selectedVar) {
   d3.csv("https://raw.githubusercontent.com/Maggieli99/Cancerstatistic/main/data/interactive.csv").then(function(data){
      x.domain(data.map(function(d) { return d.location; }))
          xAxis.call(d3.axisBottom(x))
          .selectAll("text")
          .style("text-anchor", "end")
          .attr("transform", "rotate(-75)")
          .attr("dx","-.8em")
          .attr("dy","-.5em")

          y.domain([0, 350])//d3.max(data, function(d) { return +d[selectedVar] }) ]);
          yAxis.call(d3.axisLeft(y));

          // create line

         var bars = svg.selectAll(".bar").data(data)
         
         
         bars.enter()
             .append("rect")
             .attr("class", "bar")
             .on("mouseover", function(){ d3.select(this).attr("fill", "orange");})
              .on("mouseout", function(){ d3.select(this).attr("fill", "steelblue");})

             .merge(bars)
             .transition()
             .duration(800)
                  .attr("x", function(d) {return x(d.location);})
                   .attr("y", function(d) {return y(d[selectedVar]);})
                   .attr("width", x.bandwidth()*0.95)
                   .attr("height", function(d) {return height - y(d[selectedVar])})
                   .attr("fill", "steelblue")
                   
          bars.on("mouseover", function(){ d3.select(this).attr("fill", "orange");})
              .on("mouseout", function(){ d3.select(this).attr("fill", "steelblue");});
              

    });
   }
   update('1990')




 </script>


</body>
</html>
