==> 07-conclusion.Rmd <==
# Conclusion

## Lessons

In this project, we have explore the cancer data in several interesting aspects: risks, country development, and progress. Based on our findings, we learned a lot about cancer and got the answers to our proposed questions at the very beginning. 

The relative contribution of genetic factors versus external factors has always been a dominant topic within cancer research. By analyzing the share of cancer deaths attributed to risk factors, we found that for around 1/3 of the cancer types, the share of death due to external factors are more than 50%. Risk factors contains 3 main categories: behavioral risk, environmental/occupational risk, and metabolic risk. Among these risk types, behavioral risk is at the leading position of causing deaths from cancer. And the biggest portion of behavioral risks is tobacco. We also found out that cancer favors elder men. And it is naturally connected with a deduction that elder men have high death rate from cancer because some behaviors like smoking, drinking, and been crapulent.

Over the 30 years, the life quality of people keeps improving which leads to a lower ratio of cancer death due to behavioral risk. But at the same time, people are also able to eat better and eat a lot more high-calorie and high-sweet food, which leads to higher ratio of cancer death attributed to metabolic problem such as high body-mass index and high fasting plasma glucose. Moreover, the unhealthy living style may related to the decrease in the patient's living year with disability. For example, the living-year-rate for people who live with stomach cancer do not increase; conversely, it has decreased in thirty years. This may related to unhealthy habits. So nowadays, good life style and healthy eating habit might be crucial factors to prevent having cancers. 

From data of different countries, it is not hard to find that even though the medicaltechnoloogy and cure strategy has been proven significantly, the death rate is still obviously higher in developed ccoutries. In developed countries and high-income area, although thee health cost in GDP is high, other factors including the poollution brought by the high tech can cause the death rate higher in developed coutries. Therefore, when peole focus on advanced technology, the drawback on health brought by technology  still worth people to care about. 

Overall, the death rate decreases evidently over 30 years but the death number increases. Those occurrences are both the product of the development of the world. Since the rapid improvement of advanced equipment and doctor's medical skills, cancers can be cured at the first stage. Hence, the population of the world is aging which leads to the increase in death number. Moreover, the rapid development have a positive influence on early-stage cancer patients, which makes them live in a longer life than that 30 years before. 



## Limitations

The variables in our dataset is a little bit limited since it do not have enough features to show the reason that leads to the decrease in dealth rate and the increase in death number. The factor such as world population growth or the application of some advanced equipment
can be explored and added into dataset.

## Further Direction

If we would like to further improve our project, we would like to get data set on survival rate to better analyze the progress human has made against cancer.
